Stable

+++ Deploy/Command/GitExport.php
@@ -19,7 +19,12 @@ class GitExport extends AbstractCommand
      */
     public function run()
     {
-        $workingDir = $this->config->getWorkingDirectory();
+        if ($this->get('vcs') == null) {
+            $class = get_class($this);
+            throw new \InvalidArgumentException("vcs argument is mandatory for $class command, please check you configuration");
+        }
+
+        $workingDir = $this->getWorkingDirectory();
 
         // clean working directory
         $command = sprintf("rm -rf %s", $workingDir);
@@ -30,13 +35,22 @@ class GitExport extends AbstractCommand
         $command = sprintf("mkdir -p %s", $workingDir);
         $this->shellExec($command);
 
+        // clone project
         $command = sprintf(
             "cd %s && git clone %s %s && cd %s && git checkout -f refs/tags/%s",
             $workingDir,
-            $this->config->getVcs(),
-            $this->config->getProject() . '-' . $this->input->getOption('release'),
-            $this->config->getProject() . '-' . $this->input->getOption('release'),
+            $this->get('vcs'),
+            $this->getProjectName() . '-' . $this->input->getOption('release'),
+            $this->getProjectName() . '-' . $this->input->getOption('release'),
             $this->input->getOption('release')
+        );
+
+        $this->shellExec($command);
+
+        // remove .git
+        $command = sprintf(
+            "cd %s && rm -rf .git",
+            $workingDir . '/' . $this->getProjectName() . '-' . $this->input->getOption('release')
         );
 
         $this->shellExec($command);
@@ -44,15 +58,15 @@ class GitExport extends AbstractCommand
 
     public function afterRun()
     {
+        if (!$this->isDry()) {
+            $workingDir = $this->getWorkingDirectory();
+            $projectDir = $this->getProjectName() . '-' . $this->input->getOption('release');
 
-        $workingDir = $this->config->getWorkingDirectory();
-        $projectDir = $this->config->getProject() . '-' . $this->input->getOption('release');
+            $fullDir = $workingDir . '/' . $projectDir;
 
-        $fullDir = $workingDir . '/' . $projectDir;
-
-        if (!is_dir($fullDir)) {
-            throw new \RuntimeException("git export failed, $fullDir has not been created", 400);
+            if (!is_dir($fullDir)) {
+                throw new \RuntimeException("git export failed, $fullDir has not been created", 400);
+            }
         }
-
     }
 }
